libtrx/text: move Text_Flash to libtrx
