+++ src/tr1/game/text.c
@@ -208,20 +208,6 @@ void Text_SetScale(
     text->scale.v = scale_v;
 }
 
-void Text_Flash(TEXTSTRING *const text, const bool enable, const int16_t rate)
-{
-    if (text == NULL) {
-        return;
-    }
-    if (enable) {
-        text->flags.flash = 1;
-        text->flash.rate = rate;
-        text->flash.count = rate;
-    } else {
-        text->flags.flash = 0;
-    }
-}
-
 void Text_Hide(TEXTSTRING *const text, const bool enable)
 {
     if (text == NULL) {
