+++ src/tr1/game/text.h
@@ -14,7 +14,6 @@
 RGBA_8888 Text_GetMenuColor(MENU_COLOR color);
 void Text_SetPos(TEXTSTRING *text, int16_t x, int16_t y);
 void Text_SetScale(TEXTSTRING *text, int32_t scale_h, int32_t scale_v);
-void Text_Flash(TEXTSTRING *text, bool enable, int16_t rate);
 void Text_Hide(TEXTSTRING *text, bool enable);
 void Text_AddBackground(
     TEXTSTRING *text, int16_t w, int16_t h, int16_t x, int16_t y,
