+++ src/tr2/game/text.c
@@ -68,21 +68,6 @@ void __cdecl Text_SetScale(
     text->scale.v = scale_v;
 }
 
-void __cdecl Text_Flash(
-    TEXTSTRING *const text, const int16_t enable, const int16_t rate)
-{
-    if (text == NULL) {
-        return;
-    }
-    if (enable) {
-        text->flags.flash = 1;
-        text->flash.rate = rate;
-        text->flash.count = rate;
-    } else {
-        text->flags.flash = 0;
-    }
-}
-
 void __cdecl Text_AddBackground(
     TEXTSTRING *const text, const int16_t x_size, const int16_t y_size,
     const int16_t x_off, const int16_t y_off, const int16_t z_off,
